[painter/gl] move #newTexture() to common code
It no longer depends on the GL implementation.
